// frontend/shared/types/api.ts
export interface Company {
    id: number;
    name: string;
    street_address: string;
    city: string;
    state: string;
    zip_code: string;
    country: string;
    phone_number: string;
}
// ...
export interface Token {
    id: number;
    value: string;
    status: string;
    created_on: string;
    last_used?: string | null;
    deleted_on?: string | null;
    used?: number | string;
}

export interface RequestData {
    id: number;
    data_type: string;
    http_method: string;
    server_timestamp: string;
    ip_address?: string;
    ip_source?: string;
    os?: string;
    browser?: string;
    platform?: string;
    locale?: string;
    client_time?: string;
    client_timezone?: string;
    latitude?: number;
    longitude?: number;
    location_source?: string;
}
// ...
export interface Tracking {
    id: number;
    tokens: Token[];
    token_values: string[];
    campaign: number;
    campaign_name: string;
    recipient: number | null;
    recipient_name: string;
    company: number | Company;
    count_requests: number;
    request_data?: RequestData[];
}